world/requirements.txt
numpy==2.1.0
pillow==10.4.0
tqdm==4.66.5
typing_extensions==4.12.2
